Se agrega funcionalidad para subir imagenes y conexion con cloudinary

+++ src/modules/daybook/views/EntryView.vue
@@ -9,6 +9,13 @@
             </div>
 
             <div>
+                <input type="file"
+                    ref="imageSelector"
+                    @change="onSelectedImage"
+                    v-show="false"
+                    accept="image/png, image/jpeg"
+                >
+
                 <button class="btn btn-danger mx-2"
                     v-if="entry.id"
                     @click="onDeleteEntry">
@@ -16,7 +23,8 @@
                     <i class="fa fa-trash-alt"></i>
                 </button>
 
-                <button class="btn btn-primary">
+                <button class="btn btn-primary"
+                    @click="onSelectImage">
                     Subir foto
                     <i class="fa fa-upload"></i>
                 </button>
@@ -29,10 +37,17 @@
         </div>
 
     
-        <img src="https://i.blogs.es/6f44dd/google-2015-1/450_1000.jpg" 
+        <img v-if="entry.picture && !localImage" 
+            :src="entry.picture" 
             alt="entry-picture"
             class="img-thumbnail"
         >
+        <img v-if="localImage" 
+            :src="localImage" 
+            alt="entry-picture"
+            class="img-thumbnail"
+        >
+
     </template>
 
     <Fab icon="fa-save" @on:click="saveEntry" />
@@ -41,8 +56,10 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from "vuex";
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from "../helpers/getDayMonthYear";
+import uploadImage from "../helpers/uploadImage";
 
 export default {
 
@@ -63,7 +80,8 @@ export default {
             if(this.id === 'new'){
                 return {
                     text: '',
-                    date: new Date().getTime()
+                    date: new Date().getTime(),
+                    file: null
                 }
             }
             return this.getEntryById(this.id)
@@ -83,8 +101,26 @@ export default {
         
     },
 
+    data(){
+        return {
+            localImage: null
+        }
+    },
+
     methods: {
         async saveEntry(){
+            new Swal( {
+                title: "Espere por favor",
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
+            const picture = await uploadImage(this.file)
+
+            // console.log( picture )
+
+            this.entry.picture = picture
+
             if( this.entry.id ){
                 this.updateEntry(this.entry)
             }
@@ -92,12 +128,51 @@ export default {
                 const id = await this.createEntry(this.entry)
                 this.$router.push({name: 'entry', params: {id}})
             }
+
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
+            this.file = null
         },
         async onDeleteEntry(){
-            await this.deleteEntry(this.entry.id)
-            this.$router.push({name: 'no-entry'})
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
+
+            if( isConfirmed ){
+                new Swal( {
+                    title: "Espere por favor",
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({name: 'no-entry'})
+
+                Swal.fire('Eliminado', '', 'success')
+            }
+
         },
-        ...mapActions('journal', ['updateEntry', 'createEntry', 'deleteEntry'])
+        ...mapActions('journal', ['updateEntry', 'createEntry', 'deleteEntry']),
+        onSelectedImage( event ){
+            const file = event.target.files[0]
+            if(!file) {
+                this.localImage = null
+                this.file = null
+                return
+            }
+
+            this.file = file
+
+            const fr = new FileReader()
+            fr.onload = () => this.localImage = fr.result
+            fr.readAsDataURL( file )
+            
+        },
+        onSelectImage(){
+            this.$refs.imageSelector.click()
+        }
     },
 
     created() {
